+++ kotlinx-coroutines-core/common/src/Await.kt
@@ -1,12 +1,11 @@
 /*
- * Copyright 2016-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2016-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package kotlinx.coroutines
 
 import kotlinx.atomicfu.*
 import kotlin.coroutines.*
-import kotlin.jvm.*
 
 /**
  * Awaits for completion of given deferred values without blocking a thread and resumes normally with the list of values
@@ -62,7 +61,7 @@ private class AwaitAll<T>(private val deferreds: Array<out Deferred<T>>) {
     suspend fun await(): List<T> = suspendCancellableCoroutine { cont ->
         // Intricate dance here
         // Step 1: Create nodes and install them as completion handlers, they may fire!
-        val nodes = Array<AwaitAllNode>(deferreds.size) { i ->
+        val nodes = Array(deferreds.size) { i ->
             val deferred = deferreds[i]
             deferred.start() // To properly await lazily started deferreds
             AwaitAllNode(cont, deferred).apply {
@@ -72,7 +71,7 @@ private class AwaitAll<T>(private val deferreds: Array<out Deferred<T>>) {
         val disposer = DisposeHandlersOnCancel(nodes)
         // Step 2: Set disposer to each node
         nodes.forEach { it.disposer = disposer }
-        // Here we know that if any code the nodes complete, it will dipsose the rest
+        // Here we know that if any code the nodes complete, it will dispose the rest
         // Step 3: Now we can check if continuation is complete
         if (cont.isCompleted) {
             // it is already complete while handlers were being installed -- dispose them all
@@ -94,8 +93,10 @@ private class AwaitAll<T>(private val deferreds: Array<out Deferred<T>>) {
     private inner class AwaitAllNode(private val continuation: CancellableContinuation<List<T>>, job: Job) : JobNode<Job>(job) {
         lateinit var handle: DisposableHandle
 
-        @Volatile
-        var disposer: DisposeHandlersOnCancel? = null
+        private val _disposer = atomic<DisposeHandlersOnCancel?>(null)
+        var disposer: DisposeHandlersOnCancel?
+            get() = _disposer.value
+            set(value) { _disposer.value = value }
         
         override fun invoke(cause: Throwable?) {
             if (cause != null) {
@@ -103,9 +104,8 @@ private class AwaitAll<T>(private val deferreds: Array<out Deferred<T>>) {
                 if (token != null) {
                     continuation.completeResume(token)
                     // volatile read of disposer AFTER continuation is complete
-                    val disposer = this.disposer
                     // and if disposer was already set (all handlers where already installed, then dispose them all)
-                    if (disposer != null) disposer.disposeAll()
+                    disposer?.disposeAll()
                 }
             } else if (notCompletedCount.decrementAndGet() == 0) {
                 continuation.resume(deferreds.map { it.getCompleted() })