+++ kotlinx-coroutines-core/common/src/channels/ArrayBroadcastChannel.kt
@@ -1,13 +1,13 @@
 /*
- * Copyright 2016-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2016-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package kotlinx.coroutines.channels
 
+import kotlinx.atomicfu.*
 import kotlinx.coroutines.*
 import kotlinx.coroutines.internal.*
 import kotlinx.coroutines.selects.*
-import kotlin.jvm.*
 
 /**
  * Broadcast channel with array buffer of a fixed [capacity].
@@ -44,12 +44,21 @@ internal class ArrayBroadcastChannel<E>(
 
     // head & tail are Long (64 bits) and we assume that they never wrap around
     // head, tail, and size are guarded by bufferLock
-    @Volatile
-    private var head: Long = 0 // do modulo on use of head
-    @Volatile
-    private var tail: Long = 0 // do modulo on use of tail
-    @Volatile
-    private var size: Int = 0
+
+    private val _head = atomic(0L)
+    private var head: Long // do modulo on use of head
+        get() = _head.value
+        set(value) { _head.value = value }
+
+    private val _tail = atomic(0L)
+    private var tail: Long // do modulo on use of tail
+        get() = _tail.value
+        set(value) { _tail.value = value }
+    
+    private val _size = atomic(0)
+    private var size: Int
+        get() = _size.value
+        set(value) { _size.value = value }
 
     private val subscribers = subscriberList<Subscriber<E>>()
 
@@ -105,7 +114,7 @@ internal class ArrayBroadcastChannel<E>(
             val size = this.size
             if (size >= capacity) return OFFER_FAILED
             // let's try to select sending this element to buffer
-            if (!select.trySelect(null)) { // :todo: move trySelect completion outside of lock
+            if (!select.trySelect()) { // :todo: move trySelect completion outside of lock
                 return ALREADY_SELECTED
             }
             val tail = this.tail
@@ -163,7 +172,7 @@ internal class ArrayBroadcastChannel<E>(
                     while (true) {
                         send = takeFirstSendOrPeekClosed() ?: break // when when no sender
                         if (send is Closed<*>) break // break when closed for send
-                        token = send!!.tryResumeSend(idempotent = null)
+                        token = send!!.tryResumeSend(null)
                         if (token != null) {
                             // put sent element to the buffer
                             buffer[(tail % capacity).toInt()] = (send as Send).pollResult
@@ -199,9 +208,10 @@ internal class ArrayBroadcastChannel<E>(
     ) : AbstractChannel<E>(), ReceiveChannel<E> {
         private val subLock = ReentrantLock()
 
-        @Volatile
-        @JvmField
-        var subHead: Long = 0 // guarded by subLock
+        private val _subHead = atomic(0L)
+        var subHead: Long // guarded by subLock
+            get() = _subHead.value
+            set(value) { _subHead.value = value }
 
         override val isBufferAlwaysEmpty: Boolean get() = false
         override val isBufferEmpty: Boolean get() = subHead >= broadcastChannel.tail
@@ -242,7 +252,7 @@ internal class ArrayBroadcastChannel<E>(
                     // find a receiver for an element
                     receive = takeFirstReceiveOrPeekClosed() ?: break // break when no one's receiving
                     if (receive is Closed<*>) break // noting more to do if this sub already closed
-                    token = receive.tryResumeReceive(result as E, idempotent = null)
+                    token = receive.tryResumeReceive(result as E, null)
                     if (token == null) continue // bail out here to next iteration (see for next receiver)
                     val subHead = this.subHead
                     this.subHead = subHead + 1 // retrieved element for this subscriber
@@ -296,7 +306,7 @@ internal class ArrayBroadcastChannel<E>(
                     result === POLL_FAILED -> { /* just bail out of lock */ }
                     else -> {
                         // let's try to select receiving this element from buffer
-                        if (!select.trySelect(null)) { // :todo: move trySelect completion outside of lock
+                        if (!select.trySelect()) { // :todo: move trySelect completion outside of lock
                             result = ALREADY_SELECTED
                         } else {
                             // update subHead after retrieiving element from buffer