+++ kotlinx-coroutines-core/common/src/internal/LockFreeLinkedList.common.kt
@@ -4,6 +4,8 @@
 
 package kotlinx.coroutines.internal
 
+import kotlin.jvm.*
+
 /** @suppress **This is unstable API and it is subject to change.** */
 public expect open class LockFreeLinkedListNode() {
     public val isRemoved: Boolean
@@ -49,7 +51,7 @@ public expect open class AddLastDesc<T : LockFreeLinkedListNode>(
 ) : AbstractAtomicDesc {
     val queue: LockFreeLinkedListNode
     val node: T
-    protected override fun onPrepare(affected: LockFreeLinkedListNode, next: LockFreeLinkedListNode): Any?
+    override fun finishPrepare(prepareOp: PrepareOp)
     override fun finishOnSuccess(affected: LockFreeLinkedListNode, next: LockFreeLinkedListNode)
 }
 
@@ -57,8 +59,7 @@ public expect open class AddLastDesc<T : LockFreeLinkedListNode>(
 public expect open class RemoveFirstDesc<T>(queue: LockFreeLinkedListNode): AbstractAtomicDesc {
     val queue: LockFreeLinkedListNode
     public val result: T
-    protected open fun validatePrepared(node: T): Boolean
-    protected final override fun onPrepare(affected: LockFreeLinkedListNode, next: LockFreeLinkedListNode): Any?
+    override fun finishPrepare(prepareOp: PrepareOp)
     final override fun finishOnSuccess(affected: LockFreeLinkedListNode, next: LockFreeLinkedListNode)
 }
 
@@ -68,6 +69,19 @@ public expect abstract class AbstractAtomicDesc : AtomicDesc {
     final override fun complete(op: AtomicOp<*>, failure: Any?)
     protected open fun failure(affected: LockFreeLinkedListNode): Any?
     protected open fun retry(affected: LockFreeLinkedListNode, next: Any): Boolean
-    protected abstract fun onPrepare(affected: LockFreeLinkedListNode, next: LockFreeLinkedListNode): Any? // non-null on failure
+    public abstract fun finishPrepare(prepareOp: PrepareOp) // non-null on failure
+    public open fun onPrepare(prepareOp: PrepareOp): Any? // non-null on failure
     protected abstract fun finishOnSuccess(affected: LockFreeLinkedListNode, next: LockFreeLinkedListNode)
 }
+
+/** @suppress **This is unstable API and it is subject to change.** */
+public expect class PrepareOp: OpDescriptor {
+    val affected: LockFreeLinkedListNode
+    override val atomicOp: AtomicOp<*>
+    val desc: AbstractAtomicDesc
+    fun finishPrepare()
+}
+
+@JvmField
+@SharedImmutable
+internal val REMOVE_PREPARED: Any = Symbol("REMOVE_PREPARED")