+++ reactive/kotlinx-coroutines-reactive/src/Channel.kt
@@ -5,22 +5,28 @@
 package kotlinx.coroutines.reactive
 
 import kotlinx.atomicfu.*
-import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.*
+import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.internal.*
 import org.reactivestreams.*
 
 /**
  * Subscribes to this [Publisher] and returns a channel to receive elements emitted by it.
  * The resulting channel shall be [cancelled][ReceiveChannel.cancel] to unsubscribe from this publisher.
- *
- * **Note: This API will become obsolete in future updates with introduction of lazy asynchronous streams.**
- *           See [issue #254](https://github.com/Kotlin/kotlinx.coroutines/issues/254).
- *
+
  * @param request how many items to request from publisher in advance (optional, one by default).
+ *
+ * This method is deprecated in the favor of [Flow].
+ * Instead of iterating over the resulting channel please use [collect][Flow.collect]:
+ * ```
+ * asFlow().collect { value ->
+ *     // process value
+ * }
+ * ```
  */
-@ObsoleteCoroutinesApi
-@Suppress("CONFLICTING_OVERLOADS")
+@Deprecated(
+    message = "Transforming publisher to channel is deprecated, use asFlow() instead",
+    level = DeprecationLevel.WARNING) // Will be error in 1.4
 public fun <T> Publisher<T>.openSubscription(request: Int = 1): ReceiveChannel<T> {
     val channel = SubscriptionChannel<T>(request)
     subscribe(channel)
@@ -28,7 +34,7 @@ public fun <T> Publisher<T>.openSubscription(request: Int = 1): ReceiveChannel<T
 }
 
 // Will be promoted to error in 1.3.0, removed in 1.4.0
-@Deprecated(message = "Use collect instead", level = DeprecationLevel.WARNING, replaceWith = ReplaceWith("this.collect(action)"))
+@Deprecated(message = "Use collect instead", level = DeprecationLevel.ERROR, replaceWith = ReplaceWith("this.collect(action)"))
 public suspend inline fun <T> Publisher<T>.consumeEach(action: (T) -> Unit) =
     openSubscription().consumeEach(action)
 
@@ -36,7 +42,6 @@ public suspend inline fun <T> Publisher<T>.consumeEach(action: (T) -> Unit) =
  * Subscribes to this [Publisher] and performs the specified action for each received element.
  * Cancels subscription if any exception happens during collect.
  */
-@ExperimentalCoroutinesApi // Since 1.2.1, tentatively till 1.3.0
 public suspend inline fun <T> Publisher<T>.collect(action: (T) -> Unit) =
     openSubscription().consumeEach(action)