+++ reactive/kotlinx-coroutines-rx2/src/RxChannel.kt
@@ -7,7 +7,6 @@ package kotlinx.coroutines.rx2
 import io.reactivex.*
 import io.reactivex.disposables.*
 import kotlinx.atomicfu.*
-import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.*
 import kotlinx.coroutines.internal.*
 
@@ -15,11 +14,10 @@ import kotlinx.coroutines.internal.*
  * Subscribes to this [MaybeSource] and returns a channel to receive elements emitted by it.
  * The resulting channel shall be [cancelled][ReceiveChannel.cancel] to unsubscribe from this source.
  *
- * **Note: This API will become obsolete in future updates with introduction of lazy asynchronous streams.**
- *           See [issue #254](https://github.com/Kotlin/kotlinx.coroutines/issues/254).
+ * This API is deprecated in the favour of [Flow].
+ * [MaybeSource] doesn't have a corresponding [Flow] adapter, so it should be transformed to [Observable] first.
  */
-@ObsoleteCoroutinesApi
-@Suppress("CONFLICTING_OVERLOADS")
+@Deprecated(message = "Deprecated in the favour of Flow", level = DeprecationLevel.WARNING) // Will be hidden in 1.4
 public fun <T> MaybeSource<T>.openSubscription(): ReceiveChannel<T> {
     val channel = SubscriptionChannel<T>()
     subscribe(channel)
@@ -30,11 +28,10 @@ public fun <T> MaybeSource<T>.openSubscription(): ReceiveChannel<T> {
  * Subscribes to this [ObservableSource] and returns a channel to receive elements emitted by it.
  * The resulting channel shall be [cancelled][ReceiveChannel.cancel] to unsubscribe from this source.
  *
- * **Note: This API will become obsolete in future updates with introduction of lazy asynchronous streams.**
- *           See [issue #254](https://github.com/Kotlin/kotlinx.coroutines/issues/254).
+ * This API is deprecated in the favour of [Flow].
+ * [ObservableSource] doesn't have a corresponding [Flow] adapter, so it should be transformed to [Observable] first.
  */
-@ObsoleteCoroutinesApi
-@Suppress("CONFLICTING_OVERLOADS")
+@Deprecated(message = "Deprecated in the favour of Flow", level = DeprecationLevel.WARNING) // Will be hidden in 1.4
 public fun <T> ObservableSource<T>.openSubscription(): ReceiveChannel<T> {
     val channel = SubscriptionChannel<T>()
     subscribe(channel)
@@ -42,12 +39,12 @@ public fun <T> ObservableSource<T>.openSubscription(): ReceiveChannel<T> {
 }
 
 // Will be promoted to error in 1.3.0, removed in 1.4.0
-@Deprecated(message = "Use collect instead", level = DeprecationLevel.WARNING, replaceWith = ReplaceWith("this.collect(action)"))
+@Deprecated(message = "Use collect instead", level = DeprecationLevel.ERROR, replaceWith = ReplaceWith("this.collect(action)"))
 public suspend inline fun <T> MaybeSource<T>.consumeEach(action: (T) -> Unit) =
     openSubscription().consumeEach(action)
 
 // Will be promoted to error in 1.3.0, removed in 1.4.0
-@Deprecated(message = "Use collect instead", level = DeprecationLevel.WARNING, replaceWith = ReplaceWith("this.collect(action)"))
+@Deprecated(message = "Use collect instead", level = DeprecationLevel.ERROR, replaceWith = ReplaceWith("this.collect(action)"))
 public suspend inline fun <T> ObservableSource<T>.consumeEach(action: (T) -> Unit) =
     openSubscription().consumeEach(action)
 
@@ -55,7 +52,6 @@ public suspend inline fun <T> ObservableSource<T>.consumeEach(action: (T) -> Uni
  * Subscribes to this [MaybeSource] and performs the specified action for each received element.
  * Cancels subscription if any exception happens during collect.
  */
-@ExperimentalCoroutinesApi // Since 1.2.1, tentatively till 1.3.0
 public suspend inline fun <T> MaybeSource<T>.collect(action: (T) -> Unit) =
     openSubscription().consumeEach(action)
 
@@ -63,7 +59,6 @@ public suspend inline fun <T> MaybeSource<T>.collect(action: (T) -> Unit) =
  * Subscribes to this [ObservableSource] and performs the specified action for each received element.
  * Cancels subscription if any exception happens during collect.
  */
-@ExperimentalCoroutinesApi // Since 1.2.1, tentatively till 1.3.0
 public suspend inline fun <T> ObservableSource<T>.collect(action: (T) -> Unit) =
     openSubscription().consumeEach(action)